refactor: rename handlers and helpers
